feat: Pan to sphere onClick using useBounds
When the user clicks on a sphere, the camera pans towards the object.

There are some slight issues with this, and potentially might have to refactor.

The next step is to prevent the double-pan effect once clicking on the object again, but drei has limitations with this since it's a packaged solution.

Resources:
- useBounds example: https://codesandbox.io/s/bounds-and-makedefault-rz2g0?file=/src/App.js