feat: Click to pan to sphere (fully working with position reset)
Using drei useBounds, we can pan to the objects in the scene.

The default behaviour of the helper reset the position of the camera to the entire bounds of the scene when we click off an object, but this wasn't the preferred behaviour.

I adjusted the code for onPointerMissed (click off object onto canvas) to reset the position to [0, 0, 0], which restored the camera back to the origin, but it kept looking in an odd direction.

Adding target: [0, 0, 0] maintains the current direction the user is looking, for a more fluid flow.

Resources:
- useBounds, drei documentation: https://github.com/pmndrs/drei#bounds
- useBounds example: https://codesandbox.io/s/bounds-and-makedefault-rz2g0?file=/src/App.js:2118-2293