feat: Clicking a sphere while in focus hyperlinks to original blog post
There was a huge challenge with the SelectToZoom function not letting me pass state down to the {children}.

My inital plan was to create a state (isClicked, setIsClicked) and then toggle it true when an object was selected in the scene, and false once the user clicked off, which worked.
The problem came when trying to pass the prop down to an object and not a component.

I tried to use React.cloneElement and adding the props to it, but for some reason I couldn't access the isClicked prop inside the children of that function.

The next thing I tried was to create a state inside of the Sphere itself, but there was a bug where the object in view would disappear onClick, possibly due to state re-rendering.

I then looked to useMemo, which also was a fail because I was still trying to use it like useState but prevent the re-render.

My last solution was to ignore state altogether and create a variable that held a true/false value, and create a condition inside of the onClick function that said:
  - if the value is false, then set it true -> which basically meant if the user clicked it for the first time, then remember that.
  - if the value is already true (this is second time user has clicked the object and it is now in focus), then open the original blog url in a new window, and then set it back to false.