refactor/feat: Add SASS, Enable img url extraction for text posts, HTML Overlay
SASS was pulled in from previous project and refactored for the project.
HTML was also added on top using absolute/relative positioning, currently working towards a feature that allows for the user to input their own Tumblr URL and have it extract the JSON from their blog.

[Enabling Image URLs for Text Posts]
The API returns image and text posts, but the structure for text posts are a lot different than images.
When a user reblogs an image, it is actually a "text" post and their image urls are hidden within the body of the post.

Solution
- Using regex, we can filter the string to return the first image url we find starting with the Tumblr image CDN (https://64.media.tumblr.com/) and then remove the last quotation on the end of the string using Javascript, which leaves us with the image url itself.
- In the conditional, we first sort check if the blog post is an image or a text post, then filter the data and push them into an array one by one depending on their type.
- This leaves us with the same result in the end with uniform data across all posts, making it very easy to work with when loading the image texture from the url.

Resources
- Check regex: https://regexr.com/
- Absolute / Relative Positioning (refresher): https://css-tricks.com/absolute-positioning-inside-relative-positioning/