feat: Add loading screen, Add camera reset on new data
There is now a loading state that shows while the objects are rendering into the scene.

Previously, there was a quality of life problem where if the user had scrolled around a scene, perhaps to the spheres on the edges of the grid, the camera would stay there when they would input a new tumblr.

It has been edited to reset the camera back to position [0,0,0] and target[0,0,0] on the useBounds when there is new data.

Minor Fixes
- Search input no longer has default X cancel button on certain browsers.

Resources
- r3f examples: https://docs.pmnd.rs/react-three-fiber/getting-started/examples
- r3f progressive loader example: https://codesandbox.io/s/7duy8