refactor: Fix CDN error, update from deprecated data fetching.
The previous version of r3f drei had a deprecated CDN that hosted the assets for the studio hdr lighting causing an error on load.
The newest update of drei had an import error so I had downgraded to 9.72.1 which solves the issue.

Apparently, this has been fixed as a few hours ago so the next push will be another update to drei with the latest version to see if it has been fixed.

Next.js has also deprecated getServerSideProps in favor of a fetch() function. I have refactored this to work with the current components. I did not refactor to /app/ directory.

Resources:
- pmndrs/drei issue #1500: https://github.com/pmndrs/drei/issues/1500
  - includes version to downgrade and follows import error.
- Next.js 13 data fetching: https://nextjs.org/docs/app/building-your-application/data-fetching/fetching#async-and-await-in-server-components