refactor: Add metadata back incl. fix for input zoom on mobile
Previous code in the /pages/ directory contained metadata, which includes the code to prevent zoom on mobile for the input. This has been converted to Next.js 13 next/head.

Since the font-size is under 16px, Safari will zoom in on the input and makes it almost impossible to zoom out unless the user was pinching on the input field or UI on top.

The metadata prevents the zoom and restores normal function.